Use better class name
